+++ NEWS.md
@@ -2,10 +2,11 @@
 ====================
 (2017-Nov-2)
 
-* new function "uiOutput3","textInput4" added
+* new function "uiOutput3","textInput4","textAreaInput4" added
 
 * An recursive shiny module "textFunctionInput" and "textFunction" added
 
+* Two toy shiny apps included in "inst" folder
 
 # ggplotAssist 0.1.0
 ====================